cleanup definition of publicApi
